Play around with the interpreter a bit more to add more types that can be added, and add some basic conversions.

+++ TestIL/Program.cs
@@ -6,10 +6,10 @@ namespace TestIL
     {
         static int Main(string[] args)
         {
-            int x = 5;
+            double x = 5;
             int y = 10;
 
-            return x + y;
+            return (int)(x + y);
         }
     }
 }
